refactor: split header-only drivers into source + header files
Move function implementations from .h files into corresponding .c files,
keeping only declarations, macros, and type definitions in headers.

Affected: all board/drivers/*.h and board/sys/*.h files.
No functional changes — pure structural refactor.

Closes #2171
