+++ board/drivers/power_saving.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "board/sys/sys.h"
+
+extern bool power_save_enabled;
+#ifdef ALLOW_DEBUG
+extern volatile bool stop_mode_requested;
+#endif
+
+void enable_can_transceivers(bool enabled);
+void set_power_save_state(bool enable);
+void enter_stop_mode(void);
