+++ board/sys/critical.h
@@ -1,16 +1,10 @@
+#pragma once
+
 #include "board/sys/sys.h"
 
 // ********************* Critical section helpers *********************
-uint8_t global_critical_depth = 0U;
+extern uint8_t global_critical_depth;
+extern volatile bool interrupts_enabled;
 
-static volatile bool interrupts_enabled = false;
-
-void enable_interrupts(void) {
-  interrupts_enabled = true;
-  __enable_irq();
-}
-
-void disable_interrupts(void) {
-  interrupts_enabled = false;
-  __disable_irq();
-}
+void enable_interrupts(void);
+void disable_interrupts(void);
